Split declaration validation task tests
